requirement.txt
absl-py==2.1.0
aniso8601==7.0.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.15.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
geographiclib==2.0
geonamescache==2.0.0
geopy==2.4.1
google-pasta==0.2.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.67.1
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
Jinja2==3.1.4
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
pillow==10.3.0
promise==2.3
protobuf==4.25.5
psycopg2-binary==2.9.9
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.2
requests==2.32.3
rich==13.9.4
Rx==1.6.3
scipy==1.14.1
setuptools==75.5.0
singledispatch==4.1.0
six==1.16.0
sounddevice==0.5.0
sqlparse==0.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
